+++ src/BasicManager.php
@@ -1,0 +1,36 @@
+<?php
+namespace Michaels\Manager;
+
+use Interop\Container\ContainerInterface;
+use Michaels\Manager\Contracts\ManagesItemsInterface;
+use Michaels\Manager\Traits\ManagesItemsTrait;
+
+/**
+ * Manages deeply nested, complex data.
+ *
+ * A basic manager class with no pizazz. Simply manages complex data.
+ * NOTE: this is not arrayable.
+ *
+ * @package Michaels\Manager
+ */
+class BasicManager implements
+    ManagesItemsInterface,
+    ContainerInterface
+{
+    use ManagesItemsTrait;
+
+    /**
+     * The items stored in the manager
+     * @var array $items Items governed by manager
+     */
+    protected $items;
+
+    /**
+     * Build a new manager instance
+     * @param array $items
+     */
+    public function __construct($items = [])
+    {
+        $this->initManager($items);
+    }
+}