+++ src/ConfigManager.php
@@ -6,30 +6,30 @@ use Countable;
 use Interop\Container\ContainerInterface;
 use IteratorAggregate;
 use JsonSerializable;
-use Michaels\Manager\Contracts\ChainsNestedItemsInterface;
+use Michaels\Manager\Contracts\LoadsFilesInterface;
 use Michaels\Manager\Contracts\ManagesItemsInterface;
 use Michaels\Manager\Traits\ArrayableTrait;
-use Michaels\Manager\Traits\ChainsNestedItemsTrait;
+use Michaels\Manager\Traits\LoadsFilesTrait;
 use Michaels\Manager\Traits\ManagesItemsTrait;
 
 /**
  * Manages deeply nested, complex data.
  *
- * This concrete class implements ManagesItems and ChainsNestedItems as well as
- * Container interoperability and various array functionality.
+ * This class is perfect for using Manager as a configuration bank. You can load from
+ * config files, set and retrieve, and is arrayable.
  *
  * @package Michaels\Manager
  */
-class Manager implements
+class ConfigManager implements
     ManagesItemsInterface,
-    ChainsNestedItemsInterface,
     ContainerInterface,
     ArrayAccess,
     Countable,
     IteratorAggregate,
-    JsonSerializable
+    JsonSerializable,
+    LoadsFilesInterface
 {
-    use ManagesItemsTrait, ChainsNestedItemsTrait, ArrayableTrait;
+    use ManagesItemsTrait, ArrayableTrait, LoadsFilesTrait;
 
     /**
      * The items stored in the manager