+++ src/Traits/ManagesItemsTrait.php
@@ -21,7 +21,7 @@ trait ManagesItemsTrait
      * The items stored in the manager
      * @var array $items Items governed by manager
      */
-    protected $_items;
+    protected $_items = [];
 
     /**
      * Name of the property to hold the data items. Internal use only
@@ -73,11 +73,12 @@ trait ManagesItemsTrait
      */
     public function hydrate($data, $append = false)
     {
-        if ($append === false) {
-            $this->reset($data);
-        } else {
+        if ($append) {
             $this->add($data);
+        } else {
+            $this->reset($data);
         }
+
         return $this;
     }